# pkg/gen/http/openapi3.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Runner Server",
    "description": "Credimi runner server API.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:80",
      "description": "Default server for runner-server"
    }
  ],
  "paths": {
    "/fetch-apk-and-action": {
      "post": {
        "tags": ["runner"],
        "summary": "fetch_apk_and_action runner",
        "operationId": "runner#fetch_apk_and_action",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FetchApkAndActionRequestBody"
              },
              "example": {
                "action_identifier": "Ab non et temporibus expedita ipsam omnis.",
                "instance_url": "Voluptate placeat iure neque illo voluptatum quia.",
                "version_identifier": "Quis debitis."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Fetchapkandactionresult"
                },
                "example": {
                  "apk_path": "Earum rerum id dolores sit praesentium.",
                  "code": "Doloribus est ut dolorum sed voluptatem.",
                  "version_id": "Sed veritatis deserunt vel."
                }
              }
            }
          },
          "400": {
            "description": "bad_request: Bad Request response.",
            "content": {
              "application/vnd.goa.error": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "unauthorized: Unauthorized response.",
            "content": {
              "application/vnd.goa.error": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "internal_error: Internal Server Error response.",
            "content": {
              "application/vnd.goa.error": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "502": {
            "description": "bad_gateway: Bad Gateway response.",
            "content": {
              "application/vnd.goa.error": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/process/": {
      "post": {
        "tags": ["runner"],
        "summary": "process_start_missing runner",
        "operationId": "runner#process_start_missing",
        "responses": {
          "204": { "description": "No Content response." },
          "400": {
            "description": "bad_request: Bad Request response.",
            "content": {
              "application/vnd.goa.error": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/process/list": {
      "get": {
        "tags": ["runner"],
        "summary": "process_list runner",
        "operationId": "runner#process_list",
        "responses": {
          "200": {
            "description": "OK response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "example": "Omnis accusantium autem."
                  },
                  "example": [
                    "Ea odio omnis blanditiis esse.",
                    "Vel temporibus aut.",
                    "Quae necessitatibus necessitatibus veniam similique sit.",
                    "Recusandae expedita illo omnis est culpa quia."
                  ]
                },
                "example": [
                  "Consequatur nostrum.",
                  "Tenetur est quisquam placeat."
                ]
              }
            }
          },
          "500": {
            "description": "internal_error: Internal Server Error response.",
            "content": {
              "application/vnd.goa.error": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/process/{namespace}": {
      "post": {
        "tags": ["runner"],
        "summary": "process_start runner",
        "operationId": "runner#process_start",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Consequatur consequatur sit omnis ea."
            },
            "example": "Ad aut dolores quo minus repudiandae consequuntur."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "old_namespace": { "type": "string", "example": "Fugit ea." }
                },
                "example": { "old_namespace": "Ut facilis totam modi." }
              },
              "example": {
                "old_namespace": "Natus quia qui aut enim temporibus consequatur."
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted response.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Processstartresult" },
                "example": {
                  "namespace": "Dolores minima quia.",
                  "status": "Quis reiciendis et voluptatem."
                }
              }
            }
          },
          "400": {
            "description": "bad_request: Bad Request response.",
            "content": {
              "application/vnd.goa.error": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "internal_error: Internal Server Error response.",
            "content": {
              "application/vnd.goa.error": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/store-pipeline-result": {
      "post": {
        "tags": ["runner"],
        "summary": "store_pipeline_result runner",
        "operationId": "runner#store_pipeline_result",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorePipelineResultRequestBody"
              },
              "example": {
                "instance_url": "Similique eum perferendis amet corrupti iusto.",
                "last_frame_path": "Maxime autem rerum accusamus quidem.",
                "logcat_path": "Pariatur inventore quo itaque.",
                "run_identifier": "Nobis asperiores.",
                "runner_identifier": "Quos et ipsum autem soluta cumque sit.",
                "video_path": "Rem error."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK response.",
            "content": {
              "application/json": {
                "schema": { "example": "Vero eum quaerat." },
                "example": "Iste ipsum voluptate aperiam."
              }
            }
          },
          "400": {
            "description": "bad_request: Bad Request response.",
            "content": {
              "application/vnd.goa.error": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "unauthorized: Unauthorized response.",
            "content": {
              "application/vnd.goa.error": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "internal_error: Internal Server Error response.",
            "content": {
              "application/vnd.goa.error": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "502": {
            "description": "bad_gateway: Bad Gateway response.",
            "content": {
              "application/vnd.goa.error": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/touch-fingerprint": {
      "get": {
        "tags": ["runner"],
        "summary": "touch_fingerprint runner",
        "operationId": "runner#touch_fingerprint",
        "responses": {
          "200": {
            "description": "OK response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Touchfingerprintresult"
                },
                "example": {
                  "status": "Nemo autem laudantium provident quisquam accusamus."
                }
              }
            }
          },
          "500": {
            "description": "internal_error: Internal Server Error response.",
            "content": {
              "application/vnd.goa.error": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "APIError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error domain",
            "example": "Laboriosam voluptas dolor perspiciatis."
          },
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "Quisquam quae a natus sint laboriosam qui."
          },
          "reason": {
            "type": "string",
            "description": "Error reason",
            "example": "Et odit quia ex provident dolores necessitatibus."
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code",
            "example": 1985088288595309210,
            "format": "int64"
          }
        },
        "example": {
          "error": "Aliquam ea.",
          "message": "Sint et dolores.",
          "reason": "Earum maxime voluptatem aliquid sint quas voluptatibus.",
          "status": 7296401557710804209
        },
        "required": ["status", "error", "reason", "message"]
      },
      "Error": {
        "type": "object",
        "properties": {
          "fault": {
            "type": "boolean",
            "description": "Is the error a server-side fault?",
            "example": true
          },
          "id": {
            "type": "string",
            "description": "ID is a unique identifier for this particular occurrence of the problem.",
            "example": "123abc"
          },
          "message": {
            "type": "string",
            "description": "Message is a human-readable explanation specific to this occurrence of the problem.",
            "example": "parameter 'p' must be an integer"
          },
          "name": {
            "type": "string",
            "description": "Name is the name of this class of errors.",
            "example": "bad_request"
          },
          "temporary": {
            "type": "boolean",
            "description": "Is the error temporary?",
            "example": false
          },
          "timeout": {
            "type": "boolean",
            "description": "Is the error a timeout?",
            "example": true
          }
        },
        "example": {
          "fault": false,
          "id": "123abc",
          "message": "parameter 'p' must be an integer",
          "name": "bad_request",
          "temporary": true,
          "timeout": false
        },
        "required": ["name", "id", "message", "temporary", "timeout", "fault"]
      },
      "FetchApkAndActionRequestBody": {
        "type": "object",
        "properties": {
          "action_identifier": {
            "type": "string",
            "example": "Nam facere voluptate quia sed deleniti."
          },
          "instance_url": {
            "type": "string",
            "example": "Et ut magni molestiae."
          },
          "version_identifier": {
            "type": "string",
            "example": "Rerum vel similique quisquam sunt consequatur nisi."
          }
        },
        "example": {
          "action_identifier": "Mollitia consectetur est dolor magnam adipisci molestiae.",
          "instance_url": "Ab qui pariatur ipsam consequatur nisi.",
          "version_identifier": "Velit commodi ullam tempora aliquam voluptas."
        },
        "required": ["instance_url", "version_identifier"]
      },
      "Fetchapkandactionresult": {
        "type": "object",
        "properties": {
          "apk_path": {
            "type": "string",
            "example": "Alias ipsa rem suscipit illo esse labore."
          },
          "code": {
            "type": "string",
            "example": "Molestiae voluptas aspernatur eveniet repudiandae accusantium."
          },
          "version_id": {
            "type": "string",
            "example": "Tempora occaecati provident exercitationem qui."
          }
        },
        "example": {
          "apk_path": "Rerum saepe quia temporibus dolorum sit et.",
          "code": "Cumque mollitia ut.",
          "version_id": "Reiciendis deserunt aut."
        },
        "required": ["apk_path", "version_id"]
      },
      "Processstartresult": {
        "type": "object",
        "properties": {
          "namespace": {
            "type": "string",
            "example": "Eveniet fugiat voluptatibus voluptatem non aut."
          },
          "status": { "type": "string", "example": "Asperiores velit." }
        },
        "example": {
          "namespace": "Rerum aut similique recusandae.",
          "status": "Cupiditate dolorum consequuntur voluptatibus voluptate."
        },
        "required": ["status", "namespace"]
      },
      "StorePipelineResultRequestBody": {
        "type": "object",
        "properties": {
          "instance_url": {
            "type": "string",
            "example": "Tempora ratione praesentium quos assumenda provident tempore."
          },
          "last_frame_path": {
            "type": "string",
            "example": "Iusto aspernatur rem placeat a minus officia."
          },
          "logcat_path": {
            "type": "string",
            "example": "Animi molestiae corporis repudiandae ut id."
          },
          "run_identifier": {
            "type": "string",
            "example": "Qui sit quasi dolorum illum aliquam adipisci."
          },
          "runner_identifier": {
            "type": "string",
            "example": "Voluptatum et eos."
          },
          "video_path": { "type": "string", "example": "Ea reiciendis." }
        },
        "example": {
          "instance_url": "Possimus magnam illum est labore dolor et.",
          "last_frame_path": "Doloribus enim non doloribus distinctio esse quas.",
          "logcat_path": "Repellat laudantium voluptates.",
          "run_identifier": "Qui dolores unde.",
          "runner_identifier": "Aut dignissimos.",
          "video_path": "Nam officia qui tempora quam ut."
        },
        "required": ["instance_url", "run_identifier"]
      },
      "Touchfingerprintresult": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "Illum commodi ex asperiores laboriosam."
          }
        },
        "example": { "status": "Blanditiis ullam qui esse." },
        "required": ["status"]
      }
    }
  },
  "tags": [{ "name": "runner", "description": "Runner HTTP service." }]
}