# pkg/gen/runner/views/view.go
// Code generated by goa v3.24.3, DO NOT EDIT.
//
// runner views
//
// Command:
// $ goa gen github.com/forkbombeu/credimi-runner/pkg/design

package views

import (
	goa "goa.design/goa/v3/pkg"
)

// Processstartresult is the viewed result type that is projected based on a
// view.
type Processstartresult struct {
	// Type to project
	Projected *ProcessstartresultView
	// View to render
	View string
}

// Fetchapkandactionresult is the viewed result type that is projected based on
// a view.
type Fetchapkandactionresult struct {
	// Type to project
	Projected *FetchapkandactionresultView
	// View to render
	View string
}

// Touchfingerprintresult is the viewed result type that is projected based on
// a view.
type Touchfingerprintresult struct {
	// Type to project
	Projected *TouchfingerprintresultView
	// View to render
	View string
}

// ProcessstartresultView is a type that runs validations on a projected type.
type ProcessstartresultView struct {
	Status    *string
	Namespace *string
}

// FetchapkandactionresultView is a type that runs validations on a projected
// type.
type FetchapkandactionresultView struct {
	ApkPath   *string
	VersionID *string
	Code      *string
}

// TouchfingerprintresultView is a type that runs validations on a projected
// type.
type TouchfingerprintresultView struct {
	Status *string
}

var (
	// ProcessstartresultMap is a map indexing the attribute names of
	// Processstartresult by view name.
	ProcessstartresultMap = map[string][]string{
		"default": {
			"status",
			"namespace",
		},
	}
	// FetchapkandactionresultMap is a map indexing the attribute names of
	// Fetchapkandactionresult by view name.
	FetchapkandactionresultMap = map[string][]string{
		"default": {
			"apk_path",
			"version_id",
			"code",
		},
	}
	// TouchfingerprintresultMap is a map indexing the attribute names of
	// Touchfingerprintresult by view name.
	TouchfingerprintresultMap = map[string][]string{
		"default": {
			"status",
		},
	}
)

// ValidateProcessstartresult runs the validations defined on the viewed result
// type Processstartresult.
func ValidateProcessstartresult(result *Processstartresult) (err error) {
	switch result.View {
	case "default", "":
		err = ValidateProcessstartresultView(result.Projected)
	default:
		err = goa.InvalidEnumValueError("view", result.View, []any{"default"})
	}
	return
}

// ValidateFetchapkandactionresult runs the validations defined on the viewed
// result type Fetchapkandactionresult.
func ValidateFetchapkandactionresult(result *Fetchapkandactionresult) (err error) {
	switch result.View {
	case "default", "":
		err = ValidateFetchapkandactionresultView(result.Projected)
	default:
		err = goa.InvalidEnumValueError("view", result.View, []any{"default"})
	}
	return
}

// ValidateTouchfingerprintresult runs the validations defined on the viewed
// result type Touchfingerprintresult.
func ValidateTouchfingerprintresult(result *Touchfingerprintresult) (err error) {
	switch result.View {
	case "default", "":
		err = ValidateTouchfingerprintresultView(result.Projected)
	default:
		err = goa.InvalidEnumValueError("view", result.View, []any{"default"})
	}
	return
}

// ValidateProcessstartresultView runs the validations defined on
// ProcessstartresultView using the "default" view.
func ValidateProcessstartresultView(result *ProcessstartresultView) (err error) {
	if result.Status == nil {
		err = goa.MergeErrors(err, goa.MissingFieldError("status", "result"))
	}
	if result.Namespace == nil {
		err = goa.MergeErrors(err, goa.MissingFieldError("namespace", "result"))
	}
	return
}

// ValidateFetchapkandactionresultView runs the validations defined on
// FetchapkandactionresultView using the "default" view.
func ValidateFetchapkandactionresultView(result *FetchapkandactionresultView) (err error) {
	if result.ApkPath == nil {
		err = goa.MergeErrors(err, goa.MissingFieldError("apk_path", "result"))
	}
	if result.VersionID == nil {
		err = goa.MergeErrors(err, goa.MissingFieldError("version_id", "result"))
	}
	return
}

// ValidateTouchfingerprintresultView runs the validations defined on
// TouchfingerprintresultView using the "default" view.
func ValidateTouchfingerprintresultView(result *TouchfingerprintresultView) (err error) {
	if result.Status == nil {
		err = goa.MergeErrors(err, goa.MissingFieldError("status", "result"))
	}
	return
}